Fix long cast in VideoPreviewForm.SyncSeek for correct seek clamping
Agent-Logs-Url: https://github.com/simonedeitos/dts-ad/sessions/f985cf37-b93c-4a93-a161-98f0ecf1d7ce

+++ Forms/VideoPreviewForm.cs
@@ -138,7 +138,7 @@ namespace AirDirector.Forms
             {
                 if (_vlcMediaPlayer.Length > 0)
                 {
-                    long seekMs = Math.Max(0, Math.Min(audioMs, _vlcMediaPlayer.Length));
+                    long seekMs = Math.Max(0L, Math.Min((long)audioMs, _vlcMediaPlayer.Length));
                     _vlcMediaPlayer.Time = seekMs;
                 }
             }
